requirements.txt
Django==4.2
djangorestframework
python-decouple
djangorestframework-simplejwt
celery
redis
requests
